requirements.txt
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.13
Werkzeug==2.2.2
